+++ ChangeLog.markdown
@@ -1,5 +1,12 @@
 # Yet Another DataTables Column Filter (Yadcf) Change-log
 
+## 0.6.0
+
+* Added integration with Chosen plugin (for the select filter)
+* For the integration with Chosen: added two new attributes, 'select_type' to allow turning of simple select into chosen select 
+  and 'select_type_options' used to pass an object to the chosen constructor , don't forget to inslcude the chsoen js file and read/look at the docs/showcase
+
+
 ## 0.5.8
 
 * Added new filter type: date input, make use of the jQuery UI Datepicker widget, filter_type: "date" (https://github.com/vedmack/yadcf/issues/26)